comparasion/djangoApp/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
python-dotenv==1.0.0
django-filter==23.3
Pillow==10.1.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
